refactor: 「名前の衝突」という概念にまつわるものを、フォルダに分ける。
